Add read reg/mem to gdbserver (#7625)
* bug fixes in client, and adding kill signal for server without multiprocess support
* commented out currently unnecessary packets sent by client
* Corrections to client kill signal code
* More functionality for gdbserver. Read regs and memory

+++ shlr/gdb/src/gdbclient/responses.c
@@ -89,11 +89,11 @@ int handle_qC(libgdbr_t *g) {
 			return -1;
 		} else {
 			t1++;
-			g->pid = strtol (t1, NULL, 16);
+			g->pid = (int) strtol (t1, NULL, 16);
 			t2++;
 		}
 	}
-	g->tid = strtol (t2, NULL, 16);
+	g->tid = (int) strtol (t2, NULL, 16);
 	return send_ack (g);
 }
